feat: Redux moved to login
